extract legacy graph module
